refactor: move generator helpers to top level testing package
